fix resolve_closure_parent_params
